change everything ot int

+++ src/main/java/addToCart/impl/Customer.java
@@ -2,14 +2,14 @@ package addToCart;
 
 public class Customer {
 
-    private long id;
+    private int id;
     private String name;
     private String location;
     private String username;
     private String password;
     private String email;
 
-    public Customer(long id, String name, String location, String username, String password, String email) {
+    public Customer(int id, String name, String location, String username, String password, String email) {
         this.id = id;
         this.name = name;
         this.location = location;
