+++ src/main/java/addToCart/impl/Item.java
@@ -2,12 +2,12 @@ package addToCart.impl;
 
 public class Item {
 
-    private  long id;
+    private  int id;
     private  String name;
     private  double price;
     private  String description;
 
-    public Item(long id, String name, double price, String description) {
+    public Item(int id, String name, double price, String description) {
         this.id = id;
         this.name = name;
         this.price = price;
